pyproject.toml
[project]
name = "ig-agent"
version = "0.2.0"
description = "Instagram content generation workflow using LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "cairosvg>=2.7.1",
    "ddgs==9.0.0",
    "duckduckgo-search>=8.0.0",
    "instagrapi>=2.1.3",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.13",
    "langchain-xai>=0.2.3",
    "langgraph>=0.3.30",
    "langsmith>=0.3.31",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pycairo>=1.28.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "reportlab>=4.3.1",
    "ruff>=0.11.5",
    "streamlit>=1.39.0",
    "svglib>=1.5.1",
]

[project.scripts]
ig-agent = "cli:main"
